clean: renamed file combined class FluxConverter
