+++ python/libneo/__init__.py
@@ -13,4 +13,4 @@ from .eqdsk import *
 from .eqdsk_base import *
 from .convert_neo_in_to_fortran_namelist import *
 from .getHeaderDataVMEC import *
-from .flux_label_converter import *
+from .flux_converter import *
